// generated/graphql-storefront.ts
export type Maybe<T> = T | null
// ...
export type Location = HasMetafields &
	Node & {
		__typename?: "Location"
		/** The address of the location. */
		address: LocationAddress
		/** A globally-unique ID. */
		id: Scalars["ID"]["output"]
		/** Returns a metafield found by namespace and key. */
		metafield?: Maybe<Metafield>
		/** The metafields associated with the resource matching the supplied list of namespaces and keys. */
		metafields: Array<Maybe<Metafield>>
		/** The name of the location. */
		name: Scalars["String"]["output"]
	}
// ...
export type Metafield = Node & {
	__typename?: "Metafield"
	/** The date and time when the storefront metafield was created. */
	createdAt: Scalars["DateTime"]["output"]
	/** The description of a metafield. */
	description?: Maybe<Scalars["String"]["output"]>
	/** A globally-unique ID. */
	id: Scalars["ID"]["output"]
	/** The unique identifier for the metafield within its namespace. */
	key: Scalars["String"]["output"]
	/** The container for a group of metafields that the metafield is associated with. */
	namespace: Scalars["String"]["output"]
	/** The type of resource that the metafield is attached to. */
	parentResource: MetafieldParentResource
	/** Returns a reference object if the metafield's type is a resource reference. */
	reference?: Maybe<MetafieldReference>
	/** A list of reference objects if the metafield's type is a resource reference list. */
	references?: Maybe<MetafieldReferenceConnection>
	/**
	 * The type name of the metafield.
	 * Refer to the list of [supported types](https://shopify.dev/apps/metafields/definitions/types).
	 *
	 */
	type: Scalars["String"]["output"]
	/** The date and time when the metafield was last updated. */
	updatedAt: Scalars["DateTime"]["output"]
	/** The data stored in the metafield. Always stored as a string, regardless of the metafield's type. */
	value: Scalars["String"]["output"]
}
// ...
export type Node = {
	/** A globally-unique ID. */
	id: Scalars["ID"]["output"]
}
// ...
export type StorefrontMetafieldFragmentFragment = {
	__typename?: "Metafield"
	id: string
	namespace: string
	key: string
	type: string
	value: string
	parentResource:
		| { __typename: "Article"; handle: string; id: string }
		| { __typename: "Blog"; handle: string; id: string }
		| { __typename: "Cart" }
		| { __typename: "Collection" }
		| { __typename: "Customer" }
		| { __typename: "Location" }
		| { __typename: "Market" }
		| { __typename: "Order" }
		| { __typename: "Page"; handle: string; id: string }
		| { __typename: "Product" }
		| { __typename: "ProductVariant" }
		| { __typename: "Shop" }
}
// ...
export const StorefrontMetafieldByIdentifierArticleDocument = {
	kind: "Document",
	definitions: [
		{
			kind: "OperationDefinition",
			operation: "query",
			name: { kind: "Name", value: "storefrontMetafieldByIdentifierArticle" },
			variableDefinitions: [
				{
					kind: "VariableDefinition",
					variable: {
						kind: "Variable",
						name: { kind: "Name", value: "parentId" },
					},
					type: {
						kind: "NonNullType",
						type: { kind: "NamedType", name: { kind: "Name", value: "ID" } },
					},
				},
				{
					kind: "VariableDefinition",
					variable: {
						kind: "Variable",
						name: { kind: "Name", value: "namespace" },
					},
					type: {
						kind: "NonNullType",
						type: {
							kind: "NamedType",
							name: { kind: "Name", value: "String" },
						},
					},
				},
				{
					kind: "VariableDefinition",
					variable: { kind: "Variable", name: { kind: "Name", value: "key" } },
					type: {
						kind: "NonNullType",
						type: {
							kind: "NamedType",
							name: { kind: "Name", value: "String" },
						},
					},
				},
			],
			selectionSet: {
				kind: "SelectionSet",
				selections: [
					{
						kind: "Field",
						alias: { kind: "Name", value: "record" },
						name: { kind: "Name", value: "article" },
						arguments: [
							{
								kind: "Argument",
								name: { kind: "Name", value: "id" },
								value: {
									kind: "Variable",
									name: { kind: "Name", value: "parentId" },
								},
							},
						],
						selectionSet: {
							kind: "SelectionSet",
							selections: [
								{
									kind: "Field",
									name: { kind: "Name", value: "metafield" },
									arguments: [
										{
											kind: "Argument",
											name: { kind: "Name", value: "namespace" },
											value: {
												kind: "Variable",
												name: { kind: "Name", value: "namespace" },
											},
										},
										{
											kind: "Argument",
											name: { kind: "Name", value: "key" },
											value: {
												kind: "Variable",
												name: { kind: "Name", value: "key" },
											},
										},
									],
									selectionSet: {
										kind: "SelectionSet",
										selections: [
											{
												kind: "FragmentSpread",
												name: {
													kind: "Name",
													value: "StorefrontMetafieldFragment",
												},
											},
										],
									},
								},
							],
						},
					},
				],
			},
		},
		{
			kind: "FragmentDefinition",
			name: { kind: "Name", value: "StorefrontMetafieldFragment" },
			typeCondition: {
				kind: "NamedType",
				name: { kind: "Name", value: "Metafield" },
			},
			selectionSet: {
				kind: "SelectionSet",
				selections: [
					{ kind: "Field", name: { kind: "Name", value: "id" } },
					{ kind: "Field", name: { kind: "Name", value: "namespace" } },
					{ kind: "Field", name: { kind: "Name", value: "key" } },
					{
						kind: "Field",
						name: { kind: "Name", value: "parentResource" },
						selectionSet: {
							kind: "SelectionSet",
							selections: [
								{ kind: "Field", name: { kind: "Name", value: "__typename" } },
								{
									kind: "InlineFragment",
									typeCondition: {
										kind: "NamedType",
										name: { kind: "Name", value: "Article" },
									},
									selectionSet: {
										kind: "SelectionSet",
										selections: [
											{
												kind: "Field",
												name: { kind: "Name", value: "handle" },
											},
											{ kind: "Field", name: { kind: "Name", value: "id" } },
										],
									},
								},
								{
									kind: "InlineFragment",
									typeCondition: {
										kind: "NamedType",
										name: { kind: "Name", value: "Blog" },
									},
									selectionSet: {
										kind: "SelectionSet",
										selections: [
											{
												kind: "Field",
												name: { kind: "Name", value: "handle" },
											},
											{ kind: "Field", name: { kind: "Name", value: "id" } },
										],
									},
								},
								{
									kind: "InlineFragment",
									typeCondition: {
										kind: "NamedType",
										name: { kind: "Name", value: "Page" },
									},
									selectionSet: {
										kind: "SelectionSet",
										selections: [
											{
												kind: "Field",
												name: { kind: "Name", value: "handle" },
											},
											{ kind: "Field", name: { kind: "Name", value: "id" } },
										],
									},
								},
							],
						},
					},
					{ kind: "Field", name: { kind: "Name", value: "type" } },
					{ kind: "Field", name: { kind: "Name", value: "value" } },
				],
			},
		},
	],
} as unknown as DocumentNode<
	StorefrontMetafieldByIdentifierArticleQuery,
	StorefrontMetafieldByIdentifierArticleQueryVariables
>
export const StorefrontMetafieldByIdentifierBlogDocument = {
	kind: "Document",
	definitions: [
		{
			kind: "OperationDefinition",
			operation: "query",
			name: { kind: "Name", value: "storefrontMetafieldByIdentifierBlog" },
			variableDefinitions: [
				{
					kind: "VariableDefinition",
					variable: {
						kind: "Variable",
						name: { kind: "Name", value: "parentId" },
					},
					type: {
						kind: "NonNullType",
						type: { kind: "NamedType", name: { kind: "Name", value: "ID" } },
					},
				},
				{
					kind: "VariableDefinition",
					variable: {
						kind: "Variable",
						name: { kind: "Name", value: "namespace" },
					},
					type: {
						kind: "NonNullType",
						type: {
							kind: "NamedType",
							name: { kind: "Name", value: "String" },
						},
					},
				},
				{
					kind: "VariableDefinition",
					variable: { kind: "Variable", name: { kind: "Name", value: "key" } },
					type: {
						kind: "NonNullType",
						type: {
							kind: "NamedType",
							name: { kind: "Name", value: "String" },
						},
					},
				},
			],
			selectionSet: {
				kind: "SelectionSet",
				selections: [
					{
						kind: "Field",
						alias: { kind: "Name", value: "record" },
						name: { kind: "Name", value: "blog" },
						arguments: [
							{
								kind: "Argument",
								name: { kind: "Name", value: "id" },
								value: {
									kind: "Variable",
									name: { kind: "Name", value: "parentId" },
								},
							},
						],
						selectionSet: {
							kind: "SelectionSet",
							selections: [
								{
									kind: "Field",
									name: { kind: "Name", value: "metafield" },
									arguments: [
										{
											kind: "Argument",
											name: { kind: "Name", value: "namespace" },
											value: {
												kind: "Variable",
												name: { kind: "Name", value: "namespace" },
											},
										},
										{
											kind: "Argument",
											name: { kind: "Name", value: "key" },
											value: {
												kind: "Variable",
												name: { kind: "Name", value: "key" },
											},
										},
									],
									selectionSet: {
										kind: "SelectionSet",
										selections: [
											{
												kind: "FragmentSpread",
												name: {
													kind: "Name",
													value: "StorefrontMetafieldFragment",
												},
											},
										],
									},
								},
							],
						},
					},
				],
			},
		},
		{
			kind: "FragmentDefinition",
			name: { kind: "Name", value: "StorefrontMetafieldFragment" },
			typeCondition: {
				kind: "NamedType",
				name: { kind: "Name", value: "Metafield" },
			},
			selectionSet: {
				kind: "SelectionSet",
				selections: [
					{ kind: "Field", name: { kind: "Name", value: "id" } },
					{ kind: "Field", name: { kind: "Name", value: "namespace" } },
					{ kind: "Field", name: { kind: "Name", value: "key" } },
					{
						kind: "Field",
						name: { kind: "Name", value: "parentResource" },
						selectionSet: {
							kind: "SelectionSet",
							selections: [
								{ kind: "Field", name: { kind: "Name", value: "__typename" } },
								{
									kind: "InlineFragment",
									typeCondition: {
										kind: "NamedType",
										name: { kind: "Name", value: "Article" },
									},
									selectionSet: {
										kind: "SelectionSet",
										selections: [
											{
												kind: "Field",
												name: { kind: "Name", value: "handle" },
											},
											{ kind: "Field", name: { kind: "Name", value: "id" } },
										],
									},
								},
								{
									kind: "InlineFragment",
									typeCondition: {
										kind: "NamedType",
										name: { kind: "Name", value: "Blog" },
									},
									selectionSet: {
										kind: "SelectionSet",
										selections: [
											{
												kind: "Field",
												name: { kind: "Name", value: "handle" },
											},
											{ kind: "Field", name: { kind: "Name", value: "id" } },
										],
									},
								},
								{
									kind: "InlineFragment",
									typeCondition: {
										kind: "NamedType",
										name: { kind: "Name", value: "Page" },
									},
									selectionSet: {
										kind: "SelectionSet",
										selections: [
											{
												kind: "Field",
												name: { kind: "Name", value: "handle" },
											},
											{ kind: "Field", name: { kind: "Name", value: "id" } },
										],
									},
								},
							],
						},
					},
					{ kind: "Field", name: { kind: "Name", value: "type" } },
					{ kind: "Field", name: { kind: "Name", value: "value" } },
				],
			},
		},
	],
} as unknown as DocumentNode<